monitor-clients/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mqtt-network-monitor"
version = "0.1.0"
description = "Plugin-based MQTT client agent for network monitoring"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0,<3.0",
    "psutil>=5.9,<7.0",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
mqtt-monitor = "mqtt_monitor.client:main"
